[FIX] management-requests-data-change - Get page. Get by id, Put: Now the information of the functionts its about the SELLER AND NOT THE PERSON
Why? From the beginning, they didn't tell me it was information about a SELLER, not a PERSON, which is what comes to mind when you hear about NORMAL information. Anyway, tomorrow I'll create the modal for the request because I realized there is no service that retrieves the seller's information through email, and well... I'll have to implement it now to finish quickly.